migrate shared errors to types pkg
